+++ main.rb
@@ -1,45 +1,18 @@
-require_relative 'app'
+require_relative 'refactor'
 
 def main
-  puts "Welcome to School library App!\n"
-  app = App.new
-  app.run
-end
-
-def list_holder(option)
-  case option
-  when 1
-    list_all_books
-  when 2
-    list_all_people
-  when 3
-    create_person
-  when 4
-    create_book
-  when 5
-    create_rental
-  when 6
-    list_all_rentals
-  end
-end
-
-def display_list
-  puts "Please choose an option by entering a number from below:\n"
-  puts '1 - List all books'
-  puts '2 - List all people'
-  puts '3 - Create a person(Student or Teacher)'
-  puts '4 - Create a book'
-  puts '5 - Create a rental'
-  puts '6 - List all rentals for a given person id'
-  puts '7 - Exit'
-
+  refactor = Refactor.new
+  puts 'Welcome to School Library App!'
+  refactor.display_list
   option = gets.chomp.to_i
-  list_holder(option)
-  if option == 7
-    puts 'Thank you for using this app'
+  case option
+  when 1..6 
+    refactor.list_holder(option)
+  when 7
+    puts 'Thank you for using this app!'
   else
-    display_list
+    puts 'Please enter a valid option'
   end
-end
+ end
 
 main
