+++ teacher.rb
@@ -3,7 +3,7 @@ require_relative 'person'
 class Teacher < Person
   attr_accessor :specialization
 
-  def initialize(age, specialization, name = 'unknown', parent_permission: true)
+  def initialize(age, specialization, name, parent_permission: true)
     super(age, name, parent_permission: parent_permission)
     @specialization = specialization
   end
